fix(stella-tui): the AGENTS page popup derives its inset, and a test says so
#5143 reported that #5051's wider prompt prefix left the page's command
popup one column left of the composer text it completes. **It had not, and
the issue's premise is wrong** — measured by rendering the page with the
menu open:

    PROBE col   4  "    /model"
    PROBE col   4  "    /models"
    PROBE col   0  ">>> /mod"

The popup's rect does start one column left, but each row is drawn
`" {name} "`, and that leading space is the fourth column. The name and the
composer text are both at column 4.

What was real is the second half of the issue: the `3` and the `6` were
literals that agreed with `PROMPT_PREFIX_W` by coincidence, and nothing
anywhere asserted the agreement, so the next change to the prefix would
break the alignment silently — exactly as #5051 was thought to have done.

So the numbers are derived (`MENU_INSET = PROMPT_PREFIX_W - 1`, the `- 1`
being the row's own leading space) and
`the_page_menu_starts_in_the_composers_text_column` pins the result. It
compares the two columns rather than pinning either, so it survives a change
to the prefix and fails only when the two stop tracking each other. Flip
checked: shifting `MENU_INSET` by one fails it with "different columns
(3 vs 4)".

`deck_render_snapshots.rs` crossed the 1500-line ceiling at 1547 with the
new test, and a first-time crossing takes no baseline entry — so the AGENTS
page's three goldens and `fixture_lead`, which nothing else calls, moved to
`tests/deck_render_snapshots/agents.rs` (1387 lines left in the parent). The
`mod` declaration carries an explicit `#[path]`: this file is a test root, so
a bare `mod agents;` would resolve to `tests/agents.rs` and cargo would build
it as a second test binary.

No test was deleted — the three are moved, bodies unchanged.

Closes #5143
